app_flask/requirements.txt
flask==3.0.3
Flask-SQLAlchemy
mysqlclient==2.2.4
Flask-Login==0.6.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
requests
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
pytest
pytest-flask
pytest-html
pytest-cov
